renamed interface directory
