+++ python/interface/qflex_circuit.py
@@ -2,9 +2,9 @@ import numpy as np
 
 import cirq
 
-import python.cirq_interface.qflex_virtual_device as qdevice
-import python.cirq_interface.qflex_order as qorder
-import python.cirq_interface.data_storage_interface as tmpi
+import python.interface.qflex_virtual_device as qdevice
+import python.interface.qflex_order as qorder
+import python.interface.data_storage_interface as tmpi
 
 import python.utils as qflexutils
