+++ python/qflex_cirq_example.py
@@ -2,11 +2,11 @@
 import sys, os
 sys.path.insert(
     1, os.path.realpath(os.path.dirname(os.path.realpath(__file__)) + '/../'))
-import python.cirq_interface.qflex_simulator as qsim
-import python.cirq_interface.qflex_virtual_device as qdevice
-import python.cirq_interface.qflex_grid as qgrid
-import python.cirq_interface.qflex_circuit as qcirc
-import python.cirq_interface.qflex_order as qorder
+import python.interface.qflex_simulator as qsim
+import python.interface.qflex_virtual_device as qdevice
+import python.interface.qflex_grid as qgrid
+import python.interface.qflex_circuit as qcirc
+import python.interface.qflex_order as qorder
 
 import python.utils as qflexutils
